// front-end/src/app/Models/Inputs.ts
export class Inputs {
    id?: any;
    moduleId?: number;
    name?: string;
    type?: string;
    sqlType?: string;
    sqlLength?: number;
    required?: boolean;
    addShow?: boolean;
    editShow?: boolean;
    editReadonly?: boolean;
    tableShow?: boolean;
    tableReadonly?: boolean;
    tableFilter?: boolean;
    tableOrder?: boolean;
}